Re-organize implementation files. The `rpcs` folder, while semantic for holding RPC implementations, is restrictive in that it does not describe implementations of other portions of Timetable, so it has simply been renamed to `timetable`.
