+++ src/timetable/visits_before.cc
