+++ src/timetable/visits_before_from_route.cc
